app/requirements.txt
streamlit>=1.20
torch>=1.13.0
torchvision
transformers>=4.20
faiss-cpu
Pillow
imagehash
piexif
python-magic
git+https://github.com/openai/CLIP.git
pyinstaller
fpdf2
